Refactored employee widgets new folder
