+++ lib/employee/hr_screens/update_details_screen.dart
@@ -1,6 +1,6 @@
 import 'package:echno_attendance/common_widgets/custom_app_bar.dart';
 import 'package:echno_attendance/constants/colors.dart';
-import 'package:echno_attendance/employee/widgets/update_details_form_container.dart';
+import 'package:echno_attendance/employee/widgets/hr_widgets/update_details_form_container.dart';
 import 'package:echno_attendance/utilities/helpers/helper_functions.dart';
 import 'package:echno_attendance/utilities/styles/padding_style.dart';
 import 'package:flutter/material.dart';
